edgeServer/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2025.4.26
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.7
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
numpy==2.0.1
orjson==3.11.3
pillow==10.4.0
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pyglet==2.0.16
pygments==2.19.1
pymongo==4.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
rich==14.0.0
rich-toolkit==0.14.6
setuptools==80.7.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.31
sqlalchemy-utc==0.14.0
starlette==0.37.2
trimesh==4.4.3
typer==0.15.4
typing-extensions==4.12.2
typing-inspection==0.4.1
ujson==5.11.0
uvicorn==0.30.3
uvloop==0.21.0
watchfiles==0.22.0
websockets==12.0
